implementing

+++ src/cuda/mygpu.h
@@ -36,8 +36,8 @@ public:
 	void clear();
 	void *allocate(size_t ss);
 	void print();
-	uint size_allocated(){
-		uint size = 0;
+	size_t size_allocated(){
+		size_t size = 0;
 		for(int i=0;i<MAX_DATA_SPACE;i++){
 			size += data_size[i];
 		}
